fix-product

+++ database/migrations/2025_02_22_004826_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->string('name',100);
             $table->text('description')->nullable();
-            $table->decimal('price', 10, 2)->default(0);
+            $table->integer('price')->default(0);
             $table->integer('stock_quantity')->default(0);
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
             $table->timestamps();
